requirements.txt
# packages in environment at C:\Users\LEE\anaconda3\envs\facenet-torch:
#
# Name                    Version                   Build  Channel
blinker                   1.6.2                    pypi_0    pypi
bzip2                     1.0.8                he774522_0  
ca-certificates           2023.05.30           haa95532_0  
certifi                   2023.7.22                pypi_0    pypi
charset-normalizer        3.2.0                    pypi_0    pypi
click                     8.1.6                    pypi_0    pypi
colorama                  0.4.6                    pypi_0    pypi
facenet-pytorch           2.5.3                    pypi_0    pypi
filelock                  3.12.2                   pypi_0    pypi
flask                     2.3.2                    pypi_0    pypi
idna                      3.4                      pypi_0    pypi
itsdangerous              2.1.2                    pypi_0    pypi
jinja2                    3.1.2                    pypi_0    pypi
keras                     2.13.1                   pypi_0    pypi
libffi                    3.4.4                hd77b12b_0  
markupsafe                2.1.3                    pypi_0    pypi
mpmath                    1.3.0                    pypi_0    pypi
mtcnn                     0.1.1                    pypi_0    pypi
networkx                  3.1                      pypi_0    pypi
numpy                     1.25.2                   pypi_0    pypi
opencv-python             4.8.0.76                 pypi_0    pypi
openssl                   3.0.10               h2bbff1b_0  
pillow                    10.0.0                   pypi_0    pypi
pip                       23.2.1          py310haa95532_0  
python                    3.10.12              he1021f5_0  
requests                  2.31.0                   pypi_0    pypi
setuptools                68.0.0          py310haa95532_0  
sqlite                    3.41.2               h2bbff1b_0  
sympy                     1.12                     pypi_0    pypi
tk                        8.6.12               h2bbff1b_0  
torch                     2.0.1                    pypi_0    pypi
torchvision               0.15.2                   pypi_0    pypi
typing-extensions         4.7.1                    pypi_0    pypi
tzdata                    2023c                h04d1e81_0  
urllib3                   2.0.4                    pypi_0    pypi
vc                        14.2                 h21ff451_1  
vs2015_runtime            14.27.29016          h5e58377_2  
werkzeug                  2.3.7                    pypi_0    pypi
wheel                     0.38.4          py310haa95532_0  
xz                        5.4.2                h8cc25b3_0  
zlib                      1.2.13               h8cc25b3_0  
